Group17 reachability parser optional start state (#59)
This pull request solves issue #69 

In query_grammar, we have changed the reachability rule, such that the start state is a optional.

We handle the start state as Box<Option<QueryExpression>> where it wasBox<QueryExpression> before.

In extract_system_rep.rs we have changed, such that if the start state is not given, it use the initial state of the system.

in extract_state.rs we also check whether the given clocks is defined in the system. Instead of make a new function to this, we use apply_constraints_to_state, which implies we have to make some changes.

TLDR; apply_constraints_to_state now returns a Result and OwnedFederation in a Result
We have implemented a better error handling for apply_constraints_to_state, so it return a Result
To do this we needed to wrap OwnedFederation into a Result, which mean that the datatype is now Result<OwnedFederation,String>. This means that we can error handle about if clocks exist in the transition system, which before was non-standard for OwnedFederations. This means that all clocks will now need to be handled correctly.
Changes to apply_constraints_to_state result in changing replace_vars to also return Result. We have use ? operator to propagate up the Err up to the correct place for error handling. If apply_constraints_to_state is called in function that does not return Result, we have instead used .unwrap().

Some codes have also been simplified