Group17 edge and transition ids generation (#92)
Adding IDs to both the Edge struct and Transition struct.

Edge IDs read directly from the JSON.
XML is not supported in this pull request, neither the ECDAR GUI or Reveaal uses any XML examples anyway.

Edge IDs are a unique string, we ensure uniqueness when saving to a component by remaking them when saved.

Transition IDs are purely used to figure out which EdgeID a given transition consists of (kind of like LocationID, which the implementation is based on)

Dont worry, 43 of the changed files is fixing old data. Old data did not generate edge IDs, now all the used test data has edge IDs.