Group 16 - User eviction from cache (#126)
he cache is now separated by user id.
Changed the cache to be an LRU (Least Recently Used) cache.
This allows users to be evicted once the cache is full. (Default right now is max 100 cached users)

The cache will automatically evict the least recently used user when inserting new users in the cache.

The RwLock has been changed to a mutex, this means only 1 thread can read it at a time. We have tested the benchmarks and the similar performance suggests that this is not an issue.

Co-authored-by: thom776g <thomas@nordfiber.dk>
Co-authored-by: Mads Mogensen <mail@madsmogensen.dk>